DistributionByRepoName moved to utils RepositoriesUtils
